Further improve unicode/character string tests
I noticed that the test string for unicode in 003_json.t was "château",
which as a latin1 valid string risks interacting with The Unicode Bug
from earlier versions of Perl. Using a string which can only be
represented with utf8 ensures that Perl's unicode systme is being tested
on older Perls. My favourite is \x{2764} also known as HEAVY BLACK HEART.

Also added the same unicode tests to the Storable and YAML test cases.
We can't test the in-DB representation match for Storable of course, but
the end to end test is still useful.