Move prereqs into correct dist.ini section
It turns out that JSON::MaybeXS, YAML and namespace::clean are needed
not only as test prerequisites, but also as prerequisites for the dist
itself.  This change updates the dist.ini accordingly meaning ensures
that the correct prereqs are set in the META.yml file.  This change also
fixes the CPANTS "prereq_matches_use" core metric issue.